Reorganize config and init files
