--- src/app/config/init.php
@@ -1,5 +1,0 @@
-<?php
-\TacoApp\CustomLoader::init();
-\Taco\Loader::init();
-\JasandPereza\Loader::init();
-\AddBySearch\Loader::init();
